requirements.txt
asgiref==3.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-admin-sortable2==1.0.2
django-ckeditor==6.1.0
django-import-export==2.7.1
django-js-asset==1.2.2
django-modeltranslation==0.18.5
django-unused-media==0.2.2
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.6
importlib-metadata==4.8.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==21.3
Pillow==8.4.0
pyparsing==3.1.1
python-monkey-business==1.0.0
pytils==0.3
pytz==2021.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
typing_extensions==4.0.1
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
